+++ src/main/java/de/ait/secondlife/controllers/CategoryController.java
@@ -6,7 +6,7 @@ import de.ait.secondlife.domain.dto.CategoryUpdateDto;
 import de.ait.secondlife.domain.dto.ResponseMessageDto;
 import de.ait.secondlife.exception_handling.dto.ValidationErrorsDto;
 import de.ait.secondlife.services.interfaces.CategoryService;
-import de.ait.secondlife.services.utilities.EntityUtilities;
+import de.ait.secondlife.services.utilities.UserCredentialsUtilities;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -29,7 +29,6 @@ import java.util.List;
 public class CategoryController {
 
     private final CategoryService service;
-    private final EntityUtilities utilities;
 
     @Operation(summary = "Get category by id", description = "Accessible to all users")
     @GetMapping("/{category-id}")
@@ -87,7 +86,7 @@ public class CategoryController {
                     content = @Content(mediaType = "application/json", schema = @Schema(implementation = ResponseMessageDto.class)))}
     )
     public ResponseEntity<CategoryDto> add(@Valid @RequestBody CategoryCreationDto dto) {
-        utilities.checkCredentialsToConnectImageToEntity(dto.getBaseNameOfImages());
+
         return ResponseEntity.status(HttpStatus.CREATED).body(service.save(dto));
     }
 
@@ -114,7 +113,6 @@ public class CategoryController {
             @Schema(description = "category id", example = "1")
             @PathVariable("category-id") Long categoryId,
             @RequestBody @Valid CategoryUpdateDto dto) {
-        utilities.checkCredentialsToConnectImageToEntity(dto.getBaseNameOfImages());
         return ResponseEntity.ok(service.update(categoryId, dto));
     }
