+++ src/main/java/de/ait/secondlife/controllers/OfferController.java
@@ -10,7 +10,7 @@ import de.ait.secondlife.domain.dto.*;
 import de.ait.secondlife.exception_handling.dto.ValidationErrorsDto;
 import de.ait.secondlife.exception_handling.exceptions.bad_request_exception.PaginationParameterIsWrongException;
 import de.ait.secondlife.services.interfaces.OfferService;
-import de.ait.secondlife.services.utilities.EntityUtilities;
+import de.ait.secondlife.services.utilities.UserCredentialsUtilities;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -37,7 +37,6 @@ import javax.security.auth.login.CredentialException;
 public class OfferController {
 
     private final OfferService service;
-    private final EntityUtilities utilities;
 
     private final String PAGE_VALUE = "0";
     private final String SIZE_VALUE = "10";
@@ -178,8 +177,7 @@ public class OfferController {
             @RequestBody
             @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Offer create DTO ")
             OfferCreationDto dto) throws CredentialException {
-        utilities.checkCredentialsToConnectImageToEntity(dto.getBaseNameOfImages());
-        return ResponseEntity.status(HttpStatus.CREATED).body(service.createOffer(dto));
+              return ResponseEntity.status(HttpStatus.CREATED).body(service.createOffer(dto));
     }
 
     @PutMapping
@@ -207,8 +205,7 @@ public class OfferController {
             @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Offer update DTO ")
             OfferUpdateDto dto
     ) throws CredentialException {
-        utilities.checkCredentialsToConnectImageToEntity(dto.getBaseNameOfImages());
-        return ResponseEntity.ok(service.updateOffer(dto));
+             return ResponseEntity.ok(service.updateOffer(dto));
     }
 
     @PatchMapping("/{id}/reject")
