--- src/main/java/de/ait/secondlife/exception_handling/exceptions/NoRightToChangeException.java
@@ -1,7 +1,0 @@
-package de.ait.secondlife.exception_handling.exceptions;
-
-public class NoRightToChangeException extends RuntimeException {
-    public NoRightToChangeException(String message) {
-        super(message);
-    }
-}
