+++ src/main/java/de/ait/secondlife/exception_handling/exceptions/NoRightsException.java
@@ -1,0 +1,7 @@
+package de.ait.secondlife.exception_handling.exceptions;
+
+public class NoRightsException extends RuntimeException {
+    public NoRightsException(String message) {
+        super(message);
+    }
+}
