+++ src/main/java/de/ait/secondlife/services/ImageServiceImpl.java
@@ -9,18 +9,23 @@ import de.ait.secondlife.domain.dto.ImageCreationDto;
 import de.ait.secondlife.domain.dto.ImagePathsResponseDto;
 import de.ait.secondlife.domain.entity.ImageEntity;
 import de.ait.secondlife.exception_handling.exceptions.bad_request_exception.*;
+import de.ait.secondlife.exception_handling.exceptions.not_found_exception.BadEntityTypeException;
 import de.ait.secondlife.exception_handling.exceptions.not_found_exception.ImagesNotFoundException;
 import de.ait.secondlife.repositories.ImageRepository;
-import de.ait.secondlife.services.interfaces.ImageService;
+import de.ait.secondlife.services.interfaces.*;
+import de.ait.secondlife.services.utilities.UserCredentialsUtilities;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.imageio.ImageIO;
+import javax.security.auth.login.CredentialException;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
@@ -37,6 +42,16 @@ public class ImageServiceImpl implements ImageService, ImageConstants {
 
     private final AmazonS3 s3Client;
     private final ImageRepository repository;
+    @Lazy
+    @Autowired
+    private  OfferService offerService;
+    @Lazy
+    @Autowired
+    private  UserService userService;
+    @Lazy
+    @Autowired
+    private  CategoryService categoryService;
+    private final UserCredentialsUtilities userCredentialsUtilities;
 
     @Value("${do.buket.name}")
     private String bucketName;
@@ -48,7 +63,17 @@ public class ImageServiceImpl implements ImageService, ImageConstants {
     private String basePath;
 
     @Override
-    public ImagePathsResponseDto saveNewImage(String entityType, Long entityId, ImageCreationDto dto, Long UserId) {
+    public ImagePathsResponseDto saveNewImage(String entityType, Long entityId, ImageCreationDto dto) {
+
+        checkEntityExists(entityType, entityId);
+        Long userId = -1L;
+        try {
+            userId = userService.getCurrentUser().getId();
+        } catch (CredentialException ignored) {
+        }
+
+        userCredentialsUtilities.checkUserCredentials(entityType, entityId);
+
         MultipartFile file = dto.getFile();
         checkFile(file);
 
@@ -63,7 +88,7 @@ public class ImageServiceImpl implements ImageService, ImageConstants {
         UUID baseName = UUID.randomUUID();
         Path path = entityId != null ?
                 Path.of(dirPrefix, entityType, entityId.toString()) :
-                Path.of(TEMP_IMAGE_DIR, UserId.toString(), baseName.toString());
+                Path.of(TEMP_IMAGE_DIR, userId.toString(), baseName.toString());
         Set<ImageEntity> savedImgEntities = new HashSet<>();
         fileSizes.forEach(e -> {
             String size = e[0] + "x" + e[1];
@@ -262,5 +287,17 @@ public class ImageServiceImpl implements ImageService, ImageConstants {
 
     private String makeFileName(String size, String baseName) {
         return String.format("%s_%s.%s", size, baseName, IMAGE);
+    }
+
+    public void checkEntityExists(String entityType, Long entityId) {
+        if (entityId == null) return;
+        CheckEntityExistsService service;
+        switch (EntityTypeWithImages.get(entityType.toLowerCase())) {
+            case OFFER -> service = offerService;
+            case USER -> service = userService;
+            case CATEGORY -> service = categoryService;
+            default -> throw new BadEntityTypeException(entityType);
+        }
+        if (!service.checkEntityExistsById(entityId)) throw new BadEntityTypeException(entityType, entityId);
     }
 }
